Game structure is changed. Debvided into separate files.
Color scheme is changed to black/white
